requirements.txt
beautifulsoup4==4.6.0
bleach==3.1.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.9
django-appconf==1.0.3
django-cogwheels==0.3
django-compressor==2.4
django-debug-toolbar==2.1
django-el-pagination==3.2.4
django-environ==0.4.5
django-libsass==0.8
django-modelcluster==5.0.1
django-taggit==1.2.0
django-treebeard==4.3.1
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
draftjs-exporter==2.1.7
html5lib==1.0.1
idna==2.8
libsass==0.19.4
Markdown==3.1.1
Pillow==6.2.2
pkg-resources==0.0.0
Pygments==2.5.2
pytz==2019.3
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.1.0
six==1.14.0
sqlparse==0.3.0
Unidecode==1.1.1
urllib3==1.25.7
wagtail==2.7.1
wagtailfontawesome==1.1.4
wagtailmenus==3.0
webencodings==0.5.1
Willow==1.3
